refactor(regrefugeefamily): refactored components and functions to different files
MemberList component moved to a different file. backendIntegrations.js was created for functions for
CRUD with backend like getMembersFromFamily and deleteRefugee.
